+++ src/components/user-reservation/confirmation/AccountDetails.jsx
@@ -1,0 +1,35 @@
+import { useSelector } from "react-redux";
+
+const AccountDetails = () => {
+  const { reservation } = useSelector((state) => state.reservation);
+
+  /* SAMPLE DATA */
+  const details = {
+    name: "Twin Peaks",
+    address: "501 Twin Peaks Blvd",
+    city: "San Francisco",
+    zip: "94131",
+    state: "CA",
+    phoneNumber: "123-456-7890",
+    username: "Bruce",
+    car: "Tesla Model 3",
+    licensePlate: "123abc",
+  };
+
+  return (
+    reservation && (
+      <div className="card card-compact w-[60%] bg-base-100 shadow-xl mb-5">
+        <div className="card-body">
+          <h2 className="card-title">Account Info</h2>
+          <div className="divider m-0 "></div>
+          <p>{details.username}</p>
+          <p>Cell: {details.phoneNumber}</p>
+          <p>Make & Model: {details.car}</p>
+          <p>License Plate: {details.licensePlate}</p>
+        </div>
+      </div>
+    )
+  );
+};
+
+export default AccountDetails;
